feat(custom-elements-lsp): pass user config into any plugins FUI-1600
