+++ packages/core/custom-elements-lsp/src/main/plugins/plugins.types.ts
@@ -12,7 +12,7 @@ export type Plugin = {
   diagnostics?: PartialDiagnosticsService[];
   metadata?: PartialMetadataService[];
 };
-export type CEPPlugin = (logger: Logger, services: Services) => Plugin;
+export type CEPPlugin = (logger: Logger, services: Services, userConfig: PluginConfig) => Plugin;
 
 export interface CEPPluginResistory {
   loadPlugins(packages: string[]): Promise<Array<Plugin>>;
@@ -21,3 +21,19 @@ export interface CEPPluginResistory {
 export interface CEPPluginService {
   loadPlugins(packages: string[]): Promise<Array<Plugin>>;
 }
+
+/**
+ * The type of the config which is defined in the tsconfig.json
+ * and is dependency injected into any plugin on init
+ */
+export type PluginConfig = {
+  srcRouteFromTSServer?: string;
+  designSystemPrefix?: string;
+  parser?: {
+    timeout?: number;
+    fastEnable?: boolean;
+    src?: string;
+    dependencies?: string[];
+  };
+  plugins?: string[];
+};
