+++ packages/core/custom-elements-lsp/src/main/plugins/repository.ts
@@ -1,11 +1,12 @@
 import { Logger } from 'typescript-template-language-service-decorator';
 import { Services } from '../utils/services.types';
-import { CEPPluginResistory, Plugin } from './plugins.types';
+import { CEPPluginResistory, Plugin, PluginConfig } from './plugins.types';
 
 export class CEPPluginRespistoryImpl implements CEPPluginResistory {
   constructor(
     private logger: Logger,
     private services: Services,
+    private userConfig: PluginConfig,
   ) {
     this.logger.log('Setting up CEP Plugin Repository');
   }
@@ -22,7 +23,7 @@ export class CEPPluginRespistoryImpl implements CEPPluginResistory {
         this.dynamicImporter(packageName).then((importedModule: any) => {
           const mPluginLoader = importedModule.default.default as unknown;
           if (typeof mPluginLoader !== 'function') throw new Error('Plugin must export a function');
-          resolve(mPluginLoader(this.logger, this.services));
+          resolve(mPluginLoader(this.logger, this.services, this.userConfig));
         });
       } catch (e) {
         this.logger.log(`Error loading plugin ${packageName} "${e}". Skipping...`);
